Move applying unique titles to MenuReplaceOperation
